Refactors the common code of test and bench targets into the solana_runtime::bpf_test_utils module.
